Separate methods into traits
#1254
